move tests to management command directory
